[CodeClean] separate util functions
Separate utility functions - header for log and utility.
